src: check header modified by ftrace or sth before install
Backoff when there's already ftrace on a function. Because if we insist on
overwritting function header, ftrace gets confused about its meta data and
may panic the kernel.

Ftrace and other similar infrastructures use call and jmp to do dynamic
patching. So we compare the first byte against 0xe8/0xe9 to see if we
should back off. For now, only x86-64 arch is supported.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>