border: fix "keep only one if both strong and weak symbol exist"
There are serval bugs in previous handling of weak symbols, e.g.

1. If two symbols with the same name are strong and weak, the weak one
   may sometimes overwrite global_fn_dict[name], but actually the strong
   one is finally linked into the vmlinux. e.g. skb_copy_bits

2. If weak symbol appears multiple times, the architecture-specific one
   overrides the general one, but plugsched didn't take care of this.
   e.g. arch_perf_update_userpage, arch_kexec_apply_relocations_add

3. If two symbols with the same name are global and local, and the global
   one appears before the local one in get_func_decl_strs, the condition
   (s in decl_strs and name in local_syms) meets in the latter iteration.
   It would yield "Attempt to redirect a repeating local symbol" error,
   which shouldn't happen. e.g. watchdog

4. include/linux/string.h and lib/string.c share some global functions. At
   first thought. Global function shouldn't be defined in header files.
   Otherwise it could be easily cause redifinition error at link stage.
   But they're defined as extern __always_inline, so it avoids redifinition
   Now it's clear, if there's a UND symbol strcpy, it should be resolved to
   lib/string.c instead of include/linux/string.h.

5. pugatory.ro is a stand-alone binary, arch/*/boot/compressed/string.c and
   drivers/firmware/efi/libstub/lib.a are linked into pugatory.ro but not
   vmlinux. And lib/string.c share some global functions with
   arch/boot/compressed/string.c. So lib/string.c is the correct file for
   these functions.

6. There are many "main" functions in stand-alone binaries, they will lead
   to global_fn_dict confliction.

Solution to 1. is to sort all occurance of one function name by their
linkage priority. And keep only the highest priority one. Here strong
symbols have higher priority than weak symbols.

Solution to 2. is to seperate architecture-specific weak symbol and general
weak symbol. Because two weak symbols' priority is according to the order
that the files appear in AR command. For example, arch/built-in.a overrides
kernel/built-in.a.

Solution to 3. is
to assert (file != global_fn_dict.get(name) and name not in local_syms) than
to assert (s in decl_strs and name not in local_syms) which is order-related

Solution to 4. is to skip header files when filling global_fn_dict

Solution to 5. is to remove arch/*/boot/compressed/*.c.boundary and
drivers/firmware/efi/libstub/*.c.boundary before analyze stage.

Solution to 6. is to whitelist out main functions

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>