src: fix the bug that wrong num_online_cpus used when sync status
In function parallel_state_check_init() we get num_online_cpus and
assign it to some counters, which will be used to sync status in the
stop_machine stage: __sync_sched_install/restore(), but we don't
hold the lock of the cpumask during this process. If a cpu is
online/offline between parallel_state_check_init() and stop_machine(),
the counters we got before will be wrong, which will cause bad rip
or hard lockup.

To solve this problem, we lock the cpumask before
parallel_state_check_init() and unlock it after stop_machine().

Fixes: f0bc472d05a9("cli:src: don't use lightweight stack check any more")
Signed-off-by: Cruz Zhao <CruzZhao@linux.alibaba.com>