border: move __schedule function outside in 3.10 config
The stack-pivot becomes complicated by the existence of a loop in
the __schedule, so it is better to move it out.

The orig __schedule of vmlinux:     The new __schedule of module:

                         jump to the new
	     nop     -------------------------->      nop
need_resched:                            need_resched:
        ^    top half                                 top half
        |
        |                                             springboard
        |                           jump to bottom half   |
        |    bottom half    <-----------------------------|
        |
        |    if conditions:
        |------- goto need_resched

If the conditions in bottom half are true, the RIP will jump to the
label in orignal __schedule. But we don't hope this. The __schedule
must start from the new __schedule.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>