border: support the kernel version 3.10 on x86_64
The first thing to note is that the 3.10 is not our primary maintenance
version. I just tested the latest kernel version of 3.10.0-1160
and 3.10.0.1160.1 on the CentOS 7.9, and other versions may have some
problems that need to make some adjustments.

The code base of 3.10 is more different from 4.19, so there are a lot
of changes that need to be made. In order not to break the generality
of other versions of the code, I move these chages to
post_extract.patch.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>
Co-developed-by: Yihao Wu <wuyihao@linux.alibaba.com>