src: fix bug __sched functions not checked
Functions like yield_to are located in .sched.text section. By default,
stack trace skips them. If there's a task executing these functions,
they will not be detected by stack safty check, which is unexpected.

This patch saves all stack frames including .sched.text functions, and
only ignores __schedule function. It's done by passing nosched=false in
older kernels and passing stack_trace_consume_entry as callback in newer
kernels.

Note that we can't use trace.skip to ignore __schedule function, because
according to the test, it's up to the compiler whether it will be on the
top of the stack. Alternatively, we assign function size to 0 before
sorting, and it will be skipped automatically when checking.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>