border: disable optimization of __schedule stack frame-pointer
The frame-pointer of __schedule function may or may not be optimized.
This optimization option can decide whether the RBP(x86) and X29(aarch64)
are used as stack frame-pointer. But plugsched requires that the stack
layout of __schedule in the kernel and scheduler.ko have the same number
and the same order of callee-saved registers.

For the Anolis cloud kernel, the optimization of frame pointer of
__schedule is disable by default now, and the scheduler module should be
disable too. So, the modification of prepare_task_switch should be
removed because it can optimize the frame-pointer.

For the other kernels, if the frame-pointer is optimized,
developers can change the code of scheduler module to make the
frame-pointer of scheduler module optimized too.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>