tools: add detection of frame-pointer of __schedule
For stack-pivot, there is a strong limit that the stack layout of
__schedule function in module must match to it in vmlinux. For the stack
layout, we are concerned with the stack frame register(rbp in x86 and x29
in aarch64), because whether this register is optimized can affect the
stack layout.

For this detection, I just match the first field of stack, because GCC
saves callee saved registers in a fixed order even if rbp or x29 has been
optimized, and on the other hand, __schedule function is complicated
enough that needs to store all of the callee saved registers.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>