border: fix sched_class renaming when they're not const
In some kernels, some sched_class variables are not defined as const.
And we failed to match the pattern. So we use a more stable one instead.
Note that we should replace both the following,

     const struct sched_class fair_sched_class;
=>   const struct sched_class shadow_fair_sched_class;

     struct sched_class fair_sched_class = {
=>   struct sched_class shadow_fair_sched_class = {

Because if we only replace the second one, the first one will be recognized
as a definition rather than a declaration, whose members are initialized as
NULL. And we can't replace "extern const struct sched_class ...", because
this breaks compilation.

Maybe we should work with formalized renaming mechanism later.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>