border: exclude not-linked weak functions from tainted functions
If weak symbols within the boundary are not linked into vmlinux, then they
are classified as fake-global, because the real global symbols are the
corresponding strong symbols that reside in other source files.

fake-global functions are added to scheduler-outsiders after inflection,
so that their definitions can be removed. Also they should be removed from
the insider set too.

This patch influences only confliction checking against kpatch.

Fixes: 259c48336f85("border: treat overriden weak symbols as outsiders")
Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>