src: clean and rebuild the dying tasks
When a task goes to dying, there has a gap between removing from
init_task list and releasing the task. In this case, we cannot find
these tasks from init_task list, so that the rebuilder cannot deal with
them including allocation and release memory.

The sched_cpu_dying gives us an idea that is using the pick_next_task to
find these tasks. So it is safe.

Co-developed-by: Cruz Zhao <CruzZhao@linux.alibaba.com>
Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>