cli: use __DISABLE_EXPORTS instead of removing lines from file
We met a problem when handling multiple lines of EXPORT_SYMBOL in macro,
e.g.,
1 #define SUBSYS(_x)						\
2 DEFINE_STATIC_KEY_TRUE(_x ## _cgrp_subsys_enabled_key);	\
3 DEFINE_STATIC_KEY_TRUE(_x ## _cgrp_subsys_on_dfl_key);	\
4 EXPORT_SYMBOL_GPL(_x ## _cgrp_subsys_enabled_key);		\
5 EXPORT_SYMBOL_GPL(_x ## _cgrp_subsys_on_dfl_key);

This is because the current logic of delete_export_symbol() checks files
from up to down, so the 4th line will not be deleted because it ends
with "\" at that time.

Although we could fix this issue by scanning it from down to up, the
whole logic is too complex. So we simply remove this logic and add
__DISABLE_EXPORTS when building.

Suggested-by: Yihao Wu <wuyihao@linux.alibaba.com>
Signed-off-by: Tianchen Ding <dtcccc@linux.alibaba.com>