config: add wake_up_idle_ht to interface for 5.10
__schedule is an interface function. And because wake_up_idle_ht is only
called by __schedule, it's classified as an insider function. However
__schedule is actually divided into two halves. The top half is an
interface function, while the bottom half is similar to an outsider
function. wake_up_idle_ht is called by the bottom half, so it's improper
to be classified as an inside function. Configuring it as an interface
function worksaround this bug.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>