src: use the right rule to reference symbols in module
For the symbol referenced, we have a rule:
If we reference module's symbols, we need to add the __mod_ prefix to
the symbol, and if reference vmlinux's symbols, we need to add __orig_
prefix to the symbol.

In main.c and sched_rebuild.c, some symbol references do not follow the
above rule. So I fixed them.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>