src: fixed a bug with __mod_ symbol reference count
In Makefile, the number of reference of __mod_ symbols will be counted.
The count1 indicates the number of __mod_ symbols to be referenced, and
the count2 represents the number of these symbols that exist.

If count1 != count2, then some __mod_ symbols have been optimized and
symbols do not exist.

But there's a bug in the way count2 counts. It uses the readelf tool to
count all the __mod_ symbols of .stub.o files, but undefined symbols are
counted as well. That's not right. For example, see following of count2:

   229: 0000000000000f30   197 FUNC    GLOBAL DEFAULT    1 update_rq_clock
   436: 00000000000056d0    87 FUNC    GLOBAL DEFAULT    1 set_rq_online
   437: 0000000000005730    87 FUNC    GLOBAL DEFAULT    1 set_rq_offline
   199: 0000000000000000     0 NOTYPE  GLOBAL DEFAULT  UND update_rq_clock
   134: 0000000000000000     0 NOTYPE  GLOBAL DEFAULT  UND update_rq_clock
   123: 0000000000000000     0 NOTYPE  GLOBAL DEFAULT  UND update_rq_clock
    71: 0000000000000000     0 NOTYPE  GLOBAL DEFAULT  UND set_rq_online
    72: 0000000000000000     0 NOTYPE  GLOBAL DEFAULT  UND set_rq_offline
   130: 0000000000000000     0 NOTYPE  GLOBAL DEFAULT  UND register_sched_domain_sysctl
   134: 0000000000000000     0 NOTYPE  GLOBAL DEFAULT  UND unregister_sched_domain_sysctl
    12: 0000000000000000    32 OBJECT  LOCAL  DEFAULT    7 schedstat_sops
    39: 0000000000000020    32 OBJECT  LOCAL  DEFAULT   20 sched_debug_sops
    40: 0000000000000040   272 OBJECT  LOCAL  DEFAULT   20 sched_feat_fops
   101: 0000000000002430  1318 FUNC    GLOBAL DEFAULT    1 register_sched_domain_sysctl
   117: 0000000000002980    29 FUNC    GLOBAL DEFAULT    1 unregister_sched_domain_sysctl

So we need to filter out these undefined symbols.

Co-developed-by: Yihao Wu <wuyihao@linux.alibaba.com>
Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>