cli: set the default value for the new configs
The olddefconfig target can set the default value for any new
configuration options that have been added to .config.

For the .config, it may add some new configs for different compilation
environment just like the CONFIG_GCC_PLUGIN, and if you do not specify
a build target, the kbuild system will ask the user for the value of
this configuration.

So I use the olddefconfig target to set these new configs to default
value.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>