tests: rebuild the container image when test
Now, the ci test will pull the plugsched-sdk released docker image, but
if we change some requirements, the test must fail and the Dockerfile
can not also be tested.

So, the container image should be built when testing. Because the podman
will ask user which registry to pull the image from, so I reconfigure the
regitstry config that pull the images from docker.io only.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>