contrib: unify the Dockerfile of x86_64 and aarch64
For Anolis7.9, the kernel is built by devtools and the default gcc
version is too low, which cause that the CONFIG_HAVE_STACKPROTECTOR
cannot be recognized and set to N automatically. So we use the
alinux3 container for arm64.

But Anolis8 is not, it use the gcc 8.5.1 to build the kernel which
can recongized the above config and set the right value. So we can
use the Anolis8 container, just like the x86_64.

This commit unifies the Dockerfile and adapts the ci test system.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>