contrib: standardized spec of scheduler package
The previous packaging process was very non-standard. First of all,
copying external files during the rpmbuild prep stage was
non-standard. The main work of the prep stage should be to extract
the tar package and apply patches. Therefore, I moved all the
externally copied files to the working directory.

Secondly, there were many external directories defined in the cli
build command, which is also non-standard. All files used in the
packaging process should be in the BUILD directory. Therefore, I
deleted all the defined macros and directly modified the _builddir
build directory, so that the rpmbuild build phase can directly
enter the scheduler working directory for execution.

Lastly, I modified the location of the RPM package. It was not a
good idea to place the RPM package in the plugsched installation
directory, so I put it in the scheduler working directory instead.

Signed-off-by: Erwei Deng <erwei@linux.alibaba.com>